+++ examples/example.rs
@@ -1,4 +1,3 @@
-
 extern crate smart_default;
 
 use smart_default::SmartDefault;
@@ -13,11 +12,13 @@ enum Foo {
         a: i32,
         b: i32,
         #[default(Some(Default::default()))]
-        c: Option<i32>
+        c: Option<i32>,
+        #[default(_code = "vec![1, 2, 3]")]
+        d: Vec<u32>,
     },
     Qux(i32),
 }
 
 fn main() {
-    assert!(Foo::default() == Foo::Baz { a: 12, b: 0, c: Some(0) });
+    assert!(Foo::default() == Foo::Baz { a: 12, b: 0, c: Some(0), d: vec![1, 2, 3] });
 }